+++ pages/_app.js
@@ -11,50 +11,6 @@ import supabase from "../utils/supabaseClient";
 import { AuthWrapper, useAuthContext } from "../context/auth";
 
 function MyApp({ Component, pageProps }) {
-  const [authenticatedState, setAuthenticatedState] =
-    useState("not-authenticated");
-
-  const router = useRouter();
-
-  useEffect(() => {
-    /* fires when a user signs in or out */
-    const { data: authListener } = supabase.auth.onAuthStateChange(
-      (event, session) => {
-        handleAuthChange(event, session);
-        if (event === "SIGNED_IN") {
-          setAuthenticatedState("authenticated");
-        }
-        if (event === "SIGNED_OUT") {
-          setAuthenticatedState("not-authenticated");
-          router.push("/login");
-        }
-      }
-    );
-
-    checkUser();
-    return () => {
-      authListener.unsubscribe();
-    };
-  }, []);
-
-  async function checkUser() {
-    /* when the component loads, checks user to show or hide Sign In link */
-    const user = await supabase.auth.user();
-    if (user) {
-      setAuthenticatedState("authenticated");
-    }
-  }
-
-  async function handleAuthChange(event, session) {
-    /* sets and removes the Supabase cookie */
-    await fetch("/api/auth", {
-      method: "POST",
-      headers: new Headers({ "Content-Type": "application/json" }),
-      credentials: "same-origin",
-      body: JSON.stringify({ event, session }),
-    });
-  }
-
   return (
     <AuthWrapper>
       <Layout>
